requirements.txt
torch>=2.0.0
torchvision
transformers>=4.32.0
peft>=0.5.0
accelerate>=0.23.0
pytesseract>=0.3.10
Pillow>=9.0.0
pandas
tqdm